ptrace: the tracee shouldn never change ctx->resume
Patchwork-id: 28246
O-Subject: [PATCH RHEL6 bz631968 2/5] ptrace: the tracee shouldn never change
	ctx->resume
Bugzilla: 631968
RH-Acked-by: Roland McGrath <roland@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=631968

ptrace_report_syscall_exit() is the only callback which changes ctx->resume.
Remove this, this is unnecessary. We know that ptrace_report_signal() will
be called before any other callback, and UTRACE_SIGNAL_HANDLER is not
possible.

With this change the tracer owns ctx->resume, it can always change it safely
even if the tracee is not stopped (implicit detach).

Signed-off-by: Oleg Nesterov <oleg@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>