tracing: do not allow llseek to set_ftrace_filter
Patchwork-id: 4470
O-Subject: [kernel team] [EMBARGOED PATCH RHEL6] BZ 631623 - CVE-2010-3079
	tracing: do not allow llseek to set_ftrace_filter
Bugzilla: 631626
RH-Acked-by: Eugene Teo <eugene@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

Bugzilla: 631623

CVE-2010-3079 kernel: ftrace NULL ptr deref [rhel-6.1]
https://bugzilla.redhat.com/show_bug.cgi?id=631623

Description:
============
Reading the file set_ftrace_filter does three things.

1) shows whether or not filters are set for the function tracer
2) shows what functions are set for the function tracer
3) shows what triggers are set on any functions

3 is independent from 1 and 2.

The way this file currently works is that it is a state machine,
and as you read it, it may change state. But this assumption breaks
when you use lseek() on the file. The state machine gets out of sync
and the t_show() may use the wrong pointer and cause a kernel oops.

Luckily, this will only kill the app that does the lseek, but the app
dies while holding a mutex. This prevents anyone else from using the
set_ftrace_filter file (or any other function tracing file for that matter).

A real fix for this is to rewrite the code, but that is too much for
a -rc release or stable. This patch simply disables llseek on the
set_ftrace_filter() file for now, and we can do the proper fix for the
next major release.

Upstream status:
================
Backported following patch:

- tracing: Do not allow llseek to set_ftrace_filter
  commit 9c55cb12c1c172e2d51e85fbb5a4796ca86b77e7
  Author: Steven Rostedt <srostedt@redhat.com>

Brew:
=====
https://brewweb.devel.redhat.com/taskinfo?taskID=2764898

Tested:
=======
tested with simple reproducer attached to the BZ

wbr,
jirka

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>