aio: check for multiplication overflow in do_io_submit
Patchwork-id: 28273
O-Subject: [rhel6 patch] aio: check for multiplication overflow in do_io_submit
Bugzilla: 629451
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Amerigo Wang <amwang@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

Hi,

This is a backport of the following upstream commit:

commit 75e1c70fc31490ef8a373ea2a4bea2524099b478
Author: Jeff Moyer <jmoyer@redhat.com>
Date:   Fri Sep 10 14:16:00 2010 -0700

It was pointed out that do_io_submit does not do proper bounds checking
on the passed in iocb array:

	if (unlikely(nr < 0))
		return -EINVAL;

	if (unlikely(!access_ok(VERIFY_READ, iocbpp, (nr*sizeof(*iocbpp)))))
		return -EFAULT;                      ^^^^^^^^^^^^^^^^^^^^

The attached patch ensures that that multiplication doesn't overflow by
limiting nr if it does.  This is an ok thing to do, as sys_io_submit is
documented as returning the number of iocbs submitted, so callers should
handle a return value of less than the 'nr' argument passed in.  If
there's a more clever way to do this, I'm all ears.

This resolves bug 629451.  Comments, as always, are appreciated.

Cheers,
Jeff

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>