dasd: allocate fallback cqr for reserve/release
Patchwork-id: 28287
O-Subject: [RHEL6.1 PATCH 5/6] [s390x] dasd: allocate fallback cqr for
	reserve/release
Bugzilla: 619515

Description
-----------
DASD reserve/release ioctls fail with return code ENOMEM.

The DASD reserve and release ioctls use the preallocated
memory pool of the respective device to build their CCW
requests. However, when the device is busy, the pool may
already be empty and the ioctl fails.
Usually this can be recovered by calling the ioctl again,
but in a situation in which we need to issue an unconditional
reserve to make a device operational again, this would be
not recoverable.

To avoid a failure due to lack of memory, preallocate enough
memory for a single reserve/release request, which can be
used if normal allocation fails.

Bugzilla
--------
BZ 619515
https://bugzilla.redhat.com/show_bug.cgi?id=619515

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.36
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=f932bcea6b7317312ab38df12a84710fc6c67a81

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>