virtio: console: Add a find_port_by_devt() function
Patchwork-id: 28309
O-Subject: [RHEL6.1 / 6.0.z PATCH 11/14] virtio: console: Add a
	find_port_by_devt() function
Bugzilla: 628805
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

To convert to using cdev as a pointer to avoid kref troubles, we have to
use a different method to get to a port from an inode than the current
container_of method.

Add find_port_by_devt() that looks up all portdevs and ports with those
portdevs to find the right port.

Signed-off-by: Amit Shah <amit.shah@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>