virtio: console: Add reference counting for port struct
Patchwork-id: 28311
O-Subject: [RHEL6.1 / 6.0.z PATCH 13/14] virtio: console: Add reference counting
	for port struct
Bugzilla: 628805
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

When a port got hot-unplugged, when a port was open, any file operation
after the unplugging resulted in a crash. This is fixed by ref-counting
the port structure, and releasing it only when the file is closed.

This splits the unplug operation in two parts: first marks the port
as unavailable, removes all the buffers in the vqs and removes the port
from the per-device list of ports. The second stage, invoked when all
references drop to zero, releases the chardev and frees all other memory.

Signed-off-by: Amit Shah <amit.shah@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>