virtio: console: Prevent userspace from submitting NULL buffers
Patchwork-id: 28318
O-Subject: [RHEL 6.1 / 6.0.z PATCH] virtio: console: Prevent userspace from
	submitting NULL buffers
Bugzilla: 635535
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

A userspace could submit a buffer with 0 length to be written to the
host.  Prevent such a situation.

This was not needed previously, but recent changes in the way write()
works exposed this condition to trigger a virtqueue event to the host,
causing a NULL buffer to be sent across.

Upstream: Posted
Bugzilla: 635535

Signed-off-by: Amit Shah <amit.shah@redhat.com>
CC: stable@kernel.org
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>