Range check cpu in blk_cpu_to_group
Patchwork-id: 28364
O-Subject: [PATCH RHEL6.1 BZ636981] block: Range check cpu in blk_cpu_to_group
Bugzilla: 636981
RH-Acked-by: David Howells <dhowells@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=636981

Description:
------------
While testing CPU DLPAR, the following problem was discovered.
We were DLPAR removing the first CPU, which in this case was
logical CPUs 0-3. CPUs 0-2 were already marked offline and
we were in the process of offlining CPU 3. After marking
the CPU inactive and offline in cpu_disable, but before the
cpu was completely idle (cpu_die), we ended up in __make_request
on CPU 3. There we looked at the topology map to see which CPU
to complete the I/O on and found no CPUs in the cpu_sibling_map.
This resulted in the block layer setting the completion cpu
to be NR_CPUS, which then caused an oops when we tried to
complete the I/O.

Fix this by sanity checking the value we return from blk_cpu_to_group
to be a valid cpu value.

Signed-off-by: Brian King <brking@linux.vnet.ibm.com>
Signed-off-by: Jens Axboe <jaxboe@fusionio.com>

RHEL Version Found:
-------------------
RHEL 6.0

Brew:
-----
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=2779050

Upstream:
---------
http://git.kernel.org/gitweb.cgi?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=be14eb619108fa8b7120eb2c42d66d5f623ae10e

Test Status:
------------

Tested by the IBM test team in our ppc64 stress test env.

---------------------------------------------------------------
Steve Best
IBM on-site partner

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>