mlx4: Add OFED-1.5.2 patch to increase log_mtts_per_seg
Patchwork-id: 28420
O-Subject: [PATCH RHEL-6] bz637284 Add OFED-1.5.2 patch to increase
	log_mtts_per_seg
Bugzilla: 637284

This patch only affects the mlx4 driver, and only if you try to pass a
previously illegal memory size to the module.
The corresponding patch for 5.5 will be posted immediately after this
one.  We need it in RHEL-6 to prevent a regression.

Please review and ack.

-----------------------------------------------------------------------------
As there are now machines with more physical memory, we need to be able to
register more memory. This patch lifts the upper limit of log_mtts_per_seg from
5 to 7, increasing the amount of memory that can be registered.

Signed-off-by: Eli Cohen <eli@dev.mellanox.co.il>
Rediffed by Jay Fenlason <fenlason@redhat.com> for RHEL-6

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>