GFS2: reserve more blocks for transactions
Patchwork-id: 28476
O-Subject: [RHEL6] GFS2: reserve more blocks for transactions (bz #637972)
Bugzilla: 637972
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

Some of the functions in GFS2 were not reserving space in the transaction for
the resource group header and the resource groups bitblocks that get added
when you do allocation. GFS2 now makes sure to reserve space for the
resource group header and either all the bitblocks in the resource group, or
one for each block that it may allocate, whichever is smaller using the new
gfs2_rg_blocks() inline function.

This fixes bz #637972
This patch is already in Steves upstream git repository
commit bf97b6734e027cc18abad420ab88f861f65d7816

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>