qeth: timeout on connection isolation configuration errors
Patchwork-id: 28515
O-Subject: [RHEL6.1 PATCH 1/1] [s390x] qeth: timeout on connection isolation
	configuration errors
Bugzilla: 635053

Description
-----------
Issuing for example "echo drop > /sys/bus/..../isolation"
on a device currently not capable to configure QDIO data
connection isolation results in a timeout and in a recovery.

In case of errors the IP Assist callback returns a non zero
value. This results in the framework to wait for a further
reply which will never come.
In case of errors the IP assist callback returns zero now.

Bugzilla
--------
BZ 635053
https://bugzilla.redhat.com/show_bug.cgi?id=635053

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.36
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=bbb822a8c032813148888fcec85e89edb17286d3

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.