sunrpc: increase MAX_HASHTABLE_BITS to 14
Patchwork-id: 28552
O-Subject: [RHEL6.1 PATCH 19/22] sunrpc: increase MAX_HASHTABLE_BITS to 14
Bugzilla: 637278
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Miquel van Smoorenburg <mikevs@xs4all.net>

The maximum size of the authcache is now set to 1024 (10 bits),
but on our server we need at least 4096 (12 bits). Increase
MAX_HASHTABLE_BITS to 14. This is a maximum of 16384 entries,
each containing a pointer (8 bytes on x86_64). This is
exactly the limit of kmalloc() (128K).

Signed-off-by: Miquel van Smoorenburg <mikevs@xs4all.net>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>