sctp: Fix out-of-bounds reading in sctp_asoc_get_hmac()
Patchwork-id: 28621
O-Subject: [RHEL6.1 patch] BZ640462 net: sctp: Fix out-of-bounds reading in
	sctp_asoc_get_hmac()
Bugzilla: 640462
CVE: CVE-2010-3705
RH-Acked-by: Petr Matousek <pmatouse@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

BZ640462
https://bugzilla.redhat.com/show_bug.cgi?id=640462

Description:
The sctp_asoc_get_hmac() function iterates through a peer's hmac_ids
array and attempts to ensure that only a supported hmac entry is
returned.  The current code fails to do this properly - if the last id
in the array is out of range (greater than SCTP_AUTH_HMAC_ID_MAX), the
id integer remains set after exiting the loop, and the address of an
out-of-bounds entry will be returned and subsequently used in the parent
function, causing potentially ugly memory corruption.  This patch resets
the id integer to 0 on encountering an invalid id so that NULL will be
returned after finishing the loop if no valid ids are found.

Upstream:
http://git.kernel.org/?p=linux/kernel/git/davem/net-2.6.git;a=commitdiff;h=51e97a12bef19b7e43199fc153cf9bd5f2140362

Brew:
https://brewweb.devel.redhat.com/taskinfo?taskID=2802749

Jirka

Signed-off-by: Jiri Pirko <jpirko@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>