fcoe: set default FIP mode as FIP_MODE_FABRIC
Patchwork-id: 28660
O-Subject: [PATCH RHEL 6.0]: fcoe: set default FIP mode as FIP_MODE_FABRIC
Bugzilla: 636233
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for bz 636233. It is supposed to be a 0day, so I was not
sure if I should put 6.1 or 6.0 in the subject.

>From upstream description:
Since sometimes current FIP_MODE_AUTO mode falls back to non-FIP
mode while DCB link still getting ready in fabric mode with
its peer switch, it falls back after few libfc flogi retries
and that is not we want while working with FIP enabled
switches in FABRIC mode, therefore sets default as FIP_MODE_FABRIC
as discussed and agreed before in this mail thread
http://www.open-fcoe.org/pipermail/devel/2010-August/010511.html

This patch is not upstream. It is being tested by Intel and EMC.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>