Xen PV-HVM: Enable xen pv hvm always for Xen HVM guests
Patchwork-id: 28671
O-Subject: [RHEL6.1 PATCH 4/4] Xen PV-HVM:  Enable xen pv hvm always for Xen
	HVM guests
Bugzilla: 632021
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

Now that the emul-unplug is in synch,
the disable is properly supported, and
the 8139/xen-netfront mac-addr conflict is resolved
in patches 1->3, can now enable xen pv-hvm always,
as the upstream kernel does, which provides
xen pv drivers to RHEL6 xen HVM guests without
any additional, admin/user-level changes at install
or post-install time.

It's a bit more involved then just removing a kernel parameter
and setting it to 1, since xen platform pci, which contains
xenbus support, is always configured for xen HVM guests.
This enables the removal of some hacks backported from RHEL5
to deal with no xenbus and crashes if a modprobe of xen-netfront
or blkfront was done, which received a cold reception in upstream
when xen_pv_hvm could be disabled (but can no longer be).
So, removing the additional checks cleans up the code
and brings it much closer to upstream.

The check and filter added to xen-blkfront for a pv-cdrom
connection was added to prevent a 300s delay waiting to
connect to a cdrom can't be connected.

This undoes the late, disable-xen-pv-hvm-by-default
patch in RHEL6.0 GA, and puts RHEL6.1 on-par with upstream.

Please review and ack.

- Don

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>