libfc: adds flogi retry in case DID is zero in RJT
Patchwork-id: 28673
O-Subject: [PATCH RHEL 6.0]: libfc: adds flogi retry in case DID is zero in RJT
Bugzilla: 633907

From: Mike Christie <mchristi@redhat.com>

This is for bz 633907. It is supposed to be a 0day, so I was not
sure if I should put 6.1 or 6.0 in the subject.

>From upstream description:
Sometimes switch in NPV mode rejects flogi request with DID
zero and in that case flogi is not tried again and port
remains offline, so this patch validates DID for non zero
along with only ACC response to allow flogi retry
for RJT with DID=0 also succeed FLOGI in next try.

This patch is not upstream. It is being tested by
Intel and EMC.

I have done some basic tests. I do not have a switch
with NPV support, so I just did some basic login tests.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>