Remove compat code for VIDIOCSMICROCODE
Patchwork-id: 4487
O-Subject: [kernel team] [PATCH RHEL 6.1] CVE-2010-2963: Remove compat code
	for VIDIOCSMICROCODE
Bugzilla: 642473
CVE: CVE-2010-2963
RH-Acked-by: Eugene Teo <eugene@redhat.com>

VIDIOCSMICROCODE is a way used by a few V4L1 drivers to load a firmware that was made
obsolete by request_firmware().

There only x86 driver that uses it is stradis driver (VIDEO_STRADIS). It is a driver from
kernel 2.2 days, whose the last patch from its author were the conversion to 2.4. The driver
were written back in 1999 and not enabled on RHEL.

The compat code were never tested (as we don't know anyone having one of this board), and, as
a security issue were discovered on it, better to just remove this beast.

BZ#642473

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>