xen: hold mm->page_table_lock in vmalloc_sync
Patchwork-id: 28764
O-Subject: [RHEL 6.1 PATCH] xen: hold mm->page_table_lock in vmalloc_sync
Bugzilla: 643371
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=643371

This patch was posted yesterday. As it is known to fix crashes
and may fix frequent crashes reported in another bug (more testing
needed to prove that), then I'm posting it immediately to try and
get it in sooner than later. It applies cleanly to RHEL. I've
tested it in a reboot loop (~100 reboots), and it's running now
for longer testing to see if that other bug has gone away.

Jeremy Fitzhardinge explains it in his commit message:

    Take mm->page_table_lock while syncing the vmalloc region.  This
    prevents a race with the Xen pagetable pin/unpin code, which
    expects that the page_table_lock is already held.  If this race
    occurs, then Xen can see an inconsistent page type (a page can
    either be read/write or a pagetable page, and pin/unpin converts
    it between them), which will cause either the pin or the
    set_p[gm]d to fail; either will crash the kernel.

    vmalloc_sync_all() should be called rarely, so this extra use of
    page_table_lock should not interfere with its normal users.

    The mm pointer is stashed in the pgd page's index field, as that
    won't be otherwise used for pgds.

    Bug reported by Ian Campbell <ian.cambell@eu.citrix.com>
    Derived from a patch by Jan Beulich <jbeulich@novell.com>

Brewing at
    http://brewweb.devel.redhat.com/brew/taskinfo?taskID=2828538

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>