KVM: Return EFAULT from kvm ioctl when guest accesses bad area
Patchwork-id: 28775
O-Subject: [RHEL6.1 PATCH 5/7 v2] KVM: Return EFAULT from kvm ioctl when guest
	accesses bad area
Bugzilla: 550938
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>

Resolves RHBZ 550938.

Currently if guest access address that belongs to memory slot but is not
backed up by page or page is read only KVM treats it like MMIO access.
Remove that capability. It was never part of the interface and should
not be relied upon.

Backport of what remained to be backported of upstream commit:
http://git.kernel.org/linus/edba23e51578f7cb6781461568489fc1825db4ac

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>