ext4, jbd2: Add barriers for file systems with exernal journals
Patchwork-id: 28822
O-Subject: [RHEL6.1 PATCH 33/83] ext4,
	jbd2: Add barriers for file systems with exernal journals
Bugzilla: 635199
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>
RH-Acked-by: Josef Bacik <josef@redhat.com>

From: Theodore Ts'o <tytso@mit.edu>

BZ 635199

Upstream commit: cc3e1bea5d87635c519da657303690f5538bb4eb

This is a bit complicated because we are trying to optimize when we
send barriers to the fs data disk.  We could just throw in an extra
barrier to the data disk whenever we send a barrier to the journal
disk, but that's not always strictly necessary.

We only need to send a barrier during a commit when there are data
blocks which are must be written out due to an inode written in
ordered mode, or if fsync() depends on the commit to force data blocks
to disk.  Finally, before we drop transactions from the beginning of
the journal during a checkpoint operation, we need to guarantee that
any blocks that were flushed out to the data disk are firmly on the
rust platter before we drop the transaction from the journal.

Thanks to Oleg Drokin for pointing out this flaw in ext3/ext4.

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>

NOTE: had to handpatch first hunk of fs/jbd2/commit.c due to rhel6.git
commit 17aab566

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>