IMA: require command line option to enabled
Patchwork-id: 28878
O-Subject: [RHEL6 PATCH] IMA: require command line option to enabled
Bugzilla: 643667
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

BZ 643667
Brew build: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=2832462

It was discovered by dumb luck, on a Fedora box, that IMA'a use of radix
trees to store information which had a very sparse key wastes a lot of
memory.  Because of the sparseness of the key we end up using a radix
node for every single inode in core.  As memory becomes full and we
flush the inode cache we flush the IMA radix tree entries as well, but
the whole time the inode was in core we had a structure which is 552
bytes long just taking up space.  (IMA uses a little other storage as
well, but this patch takes care of that too)

This patch disables IMA entirely by default and it can only be enabled
by setting ima=on on the command line.  Since we aren't storing IMA
information we aren't using the radix tree and thus we don't waste a
bunch of memory.

Signed-off-by: Kyle McMartin <kyle@redhat.com>
Signed-off-by: Mike Snitzer <snitzer@redhat.com>
Signed-off-by: Eric Paris <eparis@redhat.com>
Acked-by: Christoph Hellwig <chellwig@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>