remove false positive THP pmd_present BUG_ON
Patchwork-id: 28906
O-Subject: [RHEL6 PATCH] remove false positive THP pmd_present BUG_ON
Bugzilla: 646384
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Johannes Weiner <jweiner@redhat.com>

Subject: remove false positive THP pmd_present BUG_ON

From: Andrea Arcangeli <aarcange@redhat.com>

A former fix for split_huge_page to prevent 4k and 2M TLB entries to be mapped
simultanously (to workaround some CPU errata triggering machine checks)
introduced a new case where the huge pmd can be set to non present during the
tlb flush that converts it to a not-huge pmd while pmd_splitting and
pmd_trans_huge are both set. This opens a race for an even earlier BUG_ON in
migrate.c to trigger a false positive on a page that is being splitted and
whose tlb is being invalidated by the smp tlb flush. The reason for marking the
huge-splitting-pmd temporarly as non-present is to be able to solve the errata
without introducing a second tlb flush so it should be worth it to introduce
this non-present state for the pmd while it's in splitting state, even if it
makes the pmd_present callers more complex and it changes the semantics for
pmd_present a bit (for example pte_alloc_map already had to use pmd_none
instead of !pmd_present to be safe with THP enabled).

The mremap.c change at the end of this patch shouldn't be necessary because
mmap_sem is hold in write mode there, but it is worth converting the
!pmd_present to pmd_none too to avoid any confusion on the new pmd_present
semantics (it should be a noop for mremap.c).

The patch is pratically only removing the BUG_ON false positive and it's
putting the alloc_new_pmd in mremap in full sync with pte_alloc_map and friends
(the latter runs much more often), so it should be zero risk.

There's no risk of memory corruption generated from this bug either, because it
should be just a false positive BUG_ON.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>