zfcp: Fix reference counter for point-to-point port
Patchwork-id: 28911
O-Subject: [RHEL6.1 PATCH 1/1] [s390x] zfcp: Fix reference counter for
	point-to-point port
Bugzilla: 644005
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>

Description
-----------
When using an point-to-point connection for FCP and
running the sequence "enable FCP device", "disable FCP
device", "detach FCP device (e.g. vmcp det)", this
results in a hanging kernel thread kslowcrw.  Changing
CCW devices is not possible anymore when this has happened.

The reference counter of the point-to-point port is
wrong. Detaching the FCP device tries to delete all
attached ports, but it waits forever for the port's
reference counter to drop to 0.

Call zfcp_port_put after adding the point-to-point port.
zfcp_port_enqueue returns with a reference counter of 1
and the caller has to call zfcp_port_put when it does no
longer need to hold a reference to the port.

Bugzilla
--------
BZ 644005
https://bugzilla.redhat.com/show_bug.cgi?id=644005

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=f3450c7b917201bb49d67032e9f60d5125675d6a

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.