dasd: Fix unimplemented DIAG function
Patchwork-id: 28984
O-Subject: [RHEL6.1 PATCH 6/6] [s390x] dasd: Fix unimplemented DIAG function.
Bugzilla: 646951

Description
-----------
Using a DASD device with the DIAG discipline may lead to
a kernel panic when i.e. a reserve operation is executed
on this disk from another system.

The DASD interrupt handler is called for an unsolicited
interrupt and an unimplemented discipline function is called.

The information that can be provided by an unsolicited
interrupt are unimportant for a DASD device accessed with
the DIAG discipline. Simply ignore the interrupt for DIAG devices.

Bugzilla
--------
BZ 646951
https://bugzilla.redhat.com/show_bug.cgi?id=646951

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.37
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=56b86b615b807e043339979878a2de88f900ee4f

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>