cio: prevent kernel panic in I/O cancel function
Patchwork-id: 28989
O-Subject: [RHEL6.1 PATCH 1/1] [s390x] cio: prevent kernel panic in I/O
	cancel function
Bugzilla: 647825

Description
-----------
A kernel panic occurs with the following text:
"Can't stop I/O on subchannel" when a ccw device is set offline or
removed.

Linux function ccw_device_cancel_halt_clear is used to cancel an
outstanding I/O function when a ccw device is set offline or
removed. Depending on the state of the subchannel, the function
performs different I/O commands multiple times. If all attempts
to cancel the I/O failed, a kernel panic is triggered. When
a CLEAR SUBCHANNEL function is active a the subchannel while the
function is called the first time, a kernel panic is immediately
triggered because of a missing initialization of the retry
counter.

Add initialization of the retry counter. Also remove the kernel
panic when I/O could not be cancelled because this is a non-fatal
malfunction.

Bugzilla
--------
BZ 647825
https://bugzilla.redhat.com/show_bug.cgi?id=647825

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=376ae4752e3a387d41a2ba9c9ea45c2df625e6e4

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.