introduce GLF_QUEUED flag
Patchwork-id: 28992
O-Subject: [RHEL6.1 PATCH] [GFS2] - GFS1 vs GFS2 performance issue
Bugzilla: 629920
RH-Acked-by: David Teigland <teigland@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

bz 629920 (part 2 of 2)

This patch introduces a new glock flag called the queued flag. This is set any
time a holder is queued on a glock. It is reset upon a state change which does
not involve the UN state occurs, and only then if there are no queued holders
at the time.

If the flag is not set, then the minimum hold time will not be obeyed. Instead
the demote will be scheduled immediately. If the flag is set, then the min hold
time will be obeyed as usual.

This patch combined with the dlm patch posted by Dave Teigland earlier improves
gfs2 performance for the test case in the bz greatly.

GFS2's idea of which return codes it needs to handle was based
upon those listed in dlm.h. Those didn't cover all the possible
codes and listed some which never happen. This updates GFS2 to
handle all the codes which can actually be returned from the
DLM under various circumstances.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>
Tested-by: Abhijith Das <adas@redhat.com>
Signed-off-by: David Teigland <teigland@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>