ext4: Don't error out the fs if the user tries to make a file too big
Patchwork-id: 28997
O-Subject: [PATCH RHEL6.1] ext4: Don't error out the fs if the user tries to
	make a file too big
Bugzilla: 645824
RH-Acked-by: Rik van Riel <riel@redhat.com>

For Bug 645824 - ext4: Don't error out the fs if the user tries to make a file too big

If an ext3 fs is migrated to ext4, exceeding the old ext3 file
size limit on an ext3 file will throw a filesystem error, of the
sort it would if it hit metadata corruption.  This is clearly wrong,
as this triggers the (continue, remount-ro, or panic) error handler
behavior in the fs.  All we have to do is return -EFBIG and move on.
Trivial backport of the following upstream commit.

Thanks,
-Eric

commit 0c095c7f113e9fd05913d6e1b2cccbe356be039e
Author: Theodore Ts'o <tytso@mit.edu>
Date:   Tue Jul 27 11:56:06 2010 -0400

    ext4: Don't error out the fs if the user tries to make a file too big

    If the user attempts to make a non-extent-mapped file to be too large,
    return EFBIG, but don't call ext4_std_err() which will end up marking
    the file system as containing an error.

    Thanks to Toshiyuki Okajima-san at Fujitsu for pointing this out.

    Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>