Fix nfsv4 client lock reclaim behaviour
Patchwork-id: 29018
O-Subject: [RHEL6.1 patch BZ638269] Fix nfsv4 client lock reclaim behaviour
Bugzilla: 638269
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

RHEL 6 NFSv4 client lock reclaim does not work properly when after the server reboot, an IO operation which results in a STALE_STATEID response is performed before the RENEW call is sent to the server. This is a common condition in cases where the application is performing heavy IO operations on the NFS server.

While investigating this case, we came across another bug in the state recovery operation which resulted in reclaim thread looping in nfs4_reclaim_open_state().

The upstream discussion is at
http://thread.gmane.org/gmane.linux.nfs/35824/focus=35847

The patch below is a combination of the following 2 upstream patches.

--
b0ed9dbc24f1fd912b2dd08b995153cafc1d5b1c
NFSv4: Fix open recovery

NFSv4 open recovery is currently broken: since we do not clear the
state->flags states before attempting recovery, we end up with the
'can_open_cached()' function triggering. This again leads to no OPEN call
being put on the wire.

Reported-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
--

--
ae1007d37e00144b72906a4bdc47d517ae91bcc1
NFSv4: Don't call nfs4_state_mark_reclaim_reboot() from error handlers

In the case of a server reboot, the state recovery thread starts by calling
nfs4_state_end_reclaim_reboot() in order to avoid edge conditions when
the server reboots while the client is in the middle of recovery.

However, if the client has already marked the nfs4_state as requiring
reboot recovery, then the above behaviour will cause the recovery thread to
treat the open as if it was part of such an edge condition: the open will
be recovered as if it was part of a lease expiration (and all the locks
will be lost).
Fix is to remove the call to nfs4_state_mark_reclaim_reboot from
nfs4_async_handle_error(), and nfs4_handle_exception(). Instead we leave it
to the recovery thread to do this for us.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
--

Patch was compiled into the 2.6.32-81.el6 kernel and tested against the reproducer available in the bz 638269.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>