generic get_drvinfo() fallback in "ethtool.c"
Patchwork-id: 29020
O-Subject: [RHEL6.1 PATCH] generic get_drvinfo() fallback in "ethtool.c" (BZ
	#649248)
Bugzilla: 649248
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

This is a narrow backport of upstream commit
01414802054c382072b6cb9a1bdc6e243c74b2d5:

http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=01414802054c382072b6cb9a1bdc6e243c74b2d5#patch17

The original commit message is (reflowed):

----v----
The driver name and bus address for a net_device can normally be found
through the driver model now.  Instead of requiring drivers to provide
this information redundantly through the ethtool_ops::get_drvinfo
operation, use the driver model to do so if the driver does not define
the operation.  Since ETHTOOL_GDRVINFO no longer requires the driver to
implement any operations, do not require net_device::ethtool_ops to be
set either.

Remove implementations of get_drvinfo and ethtool_ops that provide only
this information.
----^----

Existing driver-specific get_drvinfo() implementations are not touched
by the backport, only the default implementation is added.

The RHEL6 version of ethtool_get_drvinfo() has a "code path for obsolete
hooks", that is, an "else" branch for an "if" where in upstream there is
none. Upstream rendered the controlling expression of the selection
statement stricter (by pre-AND-ing a non-NULL check for "ops"), which,
copied verbatim, makes the "else" branch present in RHEL6 more lenient
than it was before the backport.

This could have been fixed in multiple ways, the most frugal of them
being: extracting the non-NULL check for "ops" to the top, and have the
other stuff depend on it. Unfortunately, that would have broken the easy
comparability between the upstream and RHEL6 patches due to the change
in indentation depth. Hence the spirit of the upstream patch was kept,
and the "ops" non-nullity check was pushed down as deep and as
frequently as possible and needed; ie. "ops" is verified in the now more
lenient "else" branch again, twice. This yields the least elegant code
when looked at in separation, but also a more pleasantly looking patch.

Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=2865693

Tests:

  virtio guest, kvm host:

    before:

      # uname -r
      2.6.32-80.el6.x86_64

      # ethtool -i eth0
      Cannot get driver information: Operation not supported

    after:

      # uname -r
      2.6.32-80.el6.ethtool_backport.x86_64

      # ethtool -i eth0
      driver: virtio_net
      version:
      firmware-version:
      bus-info: virtio0

  pv guest, xen host:

    before:

      # uname -r
      2.6.32-80.el6.x86_64

      # ethtool -i eth0
      Cannot get driver information: Operation not supported

    after:

      # uname -r
      2.6.32-80.el6.ethtool_backport.x86_64

      # ethtool -i eth0
      driver: vif
      version:
      firmware-version:
      bus-info: vif-0

  pv on hvm guest (with vif = [ "...,type=netfront" ]), xen host:

    before:

      # uname -r
      2.6.32-80.el6.x86_64

      # ethtool -i eth0
      Cannot get driver information: Operation not supported

    after:

      # uname -r
      2.6.32-80.el6.ethtool_backport.x86_64

      # ethtool -i eth0
      driver: vif
      version:
      firmware-version:
      bus-info: vif-0

Thanks to Don, mst, Paolo and Drew for guidance.

Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>